requirements.txt
fastapi
farm-haystack[all]
stanza
uvicorn
pydantic
opensearch-py
scikit-learn
sentence-transformers
transformers